+++ src/main/java/org/sosly/villagetale/item/LedgerItem.java
@@ -173,11 +173,14 @@ public class LedgerItem extends Item {
 
         float health = villager.getHealth();
         int hunger = villager.getFoodData().getFoodLevel();
+        int physique = villager.getStats().getPhysique();
+        int endurance = villager.getStats().getEndurance();
+        int intellect = villager.getStats().getIntellect();
 
         List<ResourceLocation> knownRecipes = new ArrayList<>(villager.getRecipeKnowledge().known());
 
         SyncVillageCapability.send(serverPlayer, villageCapability, serverLevel.getServer());
-        OpenVillagerManagementScreen.send(serverPlayer, target.getId(), villager.getVillage().get(), homeZoneId, workZoneId, inventory, health, hunger, knownRecipes);
+        OpenVillagerManagementScreen.send(serverPlayer, target.getId(), villager.getVillage().get(), homeZoneId, workZoneId, inventory, health, hunger, physique, endurance, intellect, knownRecipes);
         return InteractionResult.CONSUME;
     }